// src/components/document-shelf/document-shelf-component-mutation-type.ts
enum DOCUMENT_SHELF_COMPONENT_MUTATION_TYPE {
    NOOP = 'noop',
    SET_DOCUMENT_FOCUSED_INDEX = "setDocumentFocusedIndex",
    SET_DOCUMENT_LIST_DISPLAY_OPTION = 'setDocumentListDisplayOption',
    SET_DOCUMENT_SORT_OPTION_SELECTED = 'setDocumentSortOptionSelected',
    SET_DOCUMENTS = 'setDocuments',
    SORT_DOCUMENTS = 'sortDocuments'
}

export default DOCUMENT_SHELF_COMPONENT_MUTATION_TYPE;